* Separated companion object of EnhancedStrategyPlayer to RankValueRegistry * Changed some functions of Player and Deck to infix functions * Some other minor tweaks
