+++ src/main/kotlin/playertypes/EnhancedStrategyPlayer.kt
@@ -5,50 +5,12 @@
 
 package playertypes
 
-import NUM_DECKS
-import NUM_SUITS
-import ProbabilityQueue
+import cardutils.RankValueRegistry
 import enums.Action
-import enums.Rank
-import java.util.PriorityQueue
-import kotlin.math.exp
 
 class EnhancedStrategyPlayer(name: String) : BasicStrategyPlayer(name) {
-    companion object {
-        private val available: MutableMap<Int, Int> = mutableMapOf()
-
-        private val numAvailableRanks: Int
-            get() = available.values.sum()
-
-        private val probabilities: ProbabilityQueue<Int>
-            get() {
-                val probQueue: ProbabilityQueue<Int> = PriorityQueue(compareByDescending { it.second })
-                for ((key, value) in available.entries)
-                    probQueue += key to exp(value.toDouble()) / available.values.sumOf { exp(it.toDouble()) }
-
-                return probQueue
-            }
-
-        private val mostLikely: Pair<Int, Double>
-            get() = probabilities.peek()
-
-        private fun fill() {
-            for (rank in Rank.values())
-                available[rank.rankValue] = (available[rank.rankValue] ?: 0) + NUM_SUITS * NUM_DECKS
-        }
-
-        fun adjust(rankValue: Int) {
-            available[rankValue] = available[rankValue]!! - 1
-            if (numAvailableRanks == 0) fill()
-        }
-    }
-
-    init {
-        fill()
-    }
-
     override fun nextAction(dealerTopCardValue: Int): Action {
-        val (rankValue, probability) = mostLikely
+        val (rankValue, probability) = RankValueRegistry.mostLikely
         return when {
             probability > 0.65 -> {
                 println("$this using probability...")
